+++ packages/expo-router/src/native-tabs/NativeBottomTabs/NativeBottomTabsNavigator.tsx
@@ -9,7 +9,7 @@ import {
   useNavigationBuilder,
   type EventMapBase,
 } from '@react-navigation/native';
-import React from 'react';
+import React, { use } from 'react';
 
 import { NativeBottomTabsRouter } from './NativeBottomTabsRouter';
 import { NativeTabsView } from './NativeTabsView';
@@ -18,12 +18,18 @@ import type { NativeTabOptions, NativeTabsProps } from './types';
 
 // In Jetpack Compose, the default back behavior is to go back to the initial route.
 const defaultBackBehavior = 'initialRoute';
+export const NativeTabsContext = React.createContext<boolean>(false);
 
 export function NativeTabsNavigator({
   children,
   backBehavior = defaultBackBehavior,
   ...rest
 }: NativeTabsProps) {
+  if (use(NativeTabsContext)) {
+    throw new Error(
+      'Nesting Native Tabs inside each other is not supported natively. Use JS tabs for nesting instead.'
+    );
+  }
   const builder = useNavigationBuilder<
     TabNavigationState<ParamListBase>,
     TabRouterOptions,
@@ -35,7 +41,11 @@ export function NativeTabsNavigator({
     backBehavior,
   });
 
-  return <NativeTabsView builder={builder} {...rest} />;
+  return (
+    <NativeTabsContext value>
+      <NativeTabsView builder={builder} {...rest} />
+    </NativeTabsContext>
+  );
 }
 
 const createNativeTabNavigator = createNavigatorFactory(NativeTabsNavigator);