+++ packages/expo-router/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### 💡 Others
 
+- Throw an error when tabs are nested ([#38651](https://github.com/expo/expo/pull/38651) by [@Ubax](https://github.com/Ubax))
+
 ## 6.0.0-beta.0 — 2025-08-13
 
 ### 🎉 New features